+++ roles/telegraf/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Toml encoder edge cases
 
 ## [2.0.8] - 2020-10-21
 ### Fixed